cnn_image/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
tensorflow==2.18.0
pillow==11.0.0
numpy==1.26.4
mlflow==3.6.0
python-multipart==0.0.20
pytest==7.4.3
httpx==0.25.2